Fix E2E FE tests on Windows to run with the new powershell script
It was identified that the python script does not work on windows to run the E2E FE tests on windows.
For that purpose, it was coded directly on the jenkins pipeline, a powershell script to run these tests.
The modifications on the tests are needed to work with the new powershell script. Some improvements were also added

Change-Id: I7a3943b6ea2ad9db59332e90cc2fa9c7755fe812