tests: fix failing MRS backend unit test
The "test_add_delete_db_object" unit test is failing due to incorrect
assertions comparing client-side case-sensitive REST object names with
actual database object names which, under certain some circumstances,
for instance, on macOS (where lower_case_table_names=2) are converted to
lowercase on lookup.

In this particular case, this affects VIEW names in the corresponding
information_schema table and PROCEDURE names (and their schema name) in
the corresponding grant table (Routine_name and Db columns in the
procs_priv table).

Since the test was failing, it was adding REST objects that did not end
up being deleted, which resulted in inconsistent state that leaked to
other tests in the suite, causing them to fail as well.

This patch addresses the issues by removing some extraneous code that
was reassigning a variable holding a db object (the VIEW) namewith its
lowercase counterpart retrieved from the I_S table and by updating a
test helper that retrieved PROCEDURE GRANTs to be a little bit less
strict and perform case-insensitive lookups.

Change-Id: I905fbc4935bd8c80b8c77144df4ef8a45869deeb