BUG#37090278 Fix default REST VIEW class name convention
By omission, when creating a REST VIEW using SQL in the MySQL Shell, if
the SDK class name is not specified, the corresponding metadata schema
record for the REST object is saved with the wrong convention.

This patch introduces the changes required to ensure the metadata schema
record for a REST VIEW is created using the correct object name.

Change-Id: Ib136e46174ef6309d064f1b4a904ee416f5ed3d6