E2E FE tests: Fix webdriver error camuflage
The FE e2e tests were encapsulating inside the beforeAll block, the creation of the webdriver instance, inside a try/catch block. When the webdriver fails to be created, a screenshot was being taken inside the catch block. Since the screenshot take is based on the webdriver instance, if it does not exist, it outputs an undefined error that camuflates the real error that is happening.
This patch also tries to debug the parallel execution of the FE tests

Change-Id: I8fa7a5982e32f25d8a5b4379dbb304120f95c42c