BUG#37154650 Grant missing invoker privileges to default user
Stored programs (procedures and functions) and views that execute in
invoker security context can perform only operations for which the
invoker has privileges. In a view, this also applies to all tables and
routines that it uses.

In MRS, if a REST view maps to a SQL view that executes in invoker
security context, the required privileges for any table or procedure
used by the view must be granted to the default MySQL Router user,
created during bootstrap, when the REST object is created.

Change-Id: I472ba46f580e7ae82389e3859892e456b749aa0a