misc: fix type errors in MRS Python SDK tests
This patch fixes all errors reported by the mypy type checker in the MRS
Python SDK test suite and tracks the corresponding mypy minimum required
version the list of development dependencies.

To accommodate those changes and to account for the actual requirements
in place, all fields in the "*Details" and "*Data" type declarations
generated for the Python SDK are now optional. This is because there is
no guarantee those fields are available in the corresponding JSON
responses sent by the MySQL Router. One such case is when the client
executes a "find*()" command and selects only specific fields to be
returned.

Change-Id: I98bd55b5d200b086a9eaba895c599707147a1cfd