Create new setup for E2E tests with TypeScript
The new setup folder contains all the necessary scripts for the new setup for typescript. The idea is to eliminate the need of more than one programming language to execute the tests, removing powershell from the loop. Typescript will replace it.
vscode-extension-tester provides an API to perform all the necessary operations (download code, chromedriver, install the extension, etc), and we are going to use with this patch.
Several other refactors were made along with the new setup.

Change-Id: Idd13e6c3847ebfbbce87377eddd52e6baea6f77c