Hide Routine result metadata in MRS TypeScript SDK
When a REST routine (Function or Procedure) is called, the response sent
by the MySQL Router can include transactional metadata like a GTID, or
in the case of Procedure result set items (if they exist) details about
the column name and type. The former is meant to be explicitly tracked
by the MRS SDK business logic, and not the application. The latter is
irrelevant to the SDK because it should already be available when the
SDK is generated in the first place, in the form of language-specific
type definitions.

This patch introduces the changes to discard metadata associated to the
result of a REST Function or Procedure from the application-level
interface objects returned by the corresponding "call()" commands in the
MRS TypeScript SDK.

Change-Id: I0d7151c2c51a3361e77ddceb356f174533927d87