Add the Start of a sandbox instance on Extension E2E tests
This patch fixes an issue found when running the tests on windows. When checking if the "mysql_rest_service_metadata" schema exists, it should connect to the port defined on the env var "MYSQL_PORT", which can be the sandbox instance port. Previously, it always connected to 3306, which lead to test failures

Change-Id: I9641f93a3aa48d0e2a7f9a167d55c35d468481d9