WL#16591 Add support for MySQL auth in the MRS TS SDK
In the MySQL REST Service, when a REST object under a given service
requires authentication to be accessed or modified, it is expected that
the service provides a corresponding REST authentication app that allows
clients to effectively do that.

In the scope of the MRS SDK, there are two different types of
authentication apps that can be supported, those based on MRS Native
authentication, where REST users are managed in the scope of the
service, and those based on MySQL Internal authentication, where users
are managed by the underlying MySQL Server.

Even though the TypeScript SDK already had limited interactive support
for MRS Native authentication apps, the API itself only materialized in
the scope of MySQL Shell GUI and not in the exported SDK files.

This patch introduces the changes required to allow applications to
authenticate in a given REST service with the MRS SDK (regardless of
where it is deployed) using both MRS Native and MySQL Internal
authentication apps.

Change-Id: I0d98c5462abfe897b63a2735e9ccdfaccdcecb8a