OCI OAuth login in the MRS landing page
This patch updates the MRS landing page logic to allow a user to
authenticate in a given REST service using OCI OAuth.

Additionally, it introduces changes to force new REST services to
support only HTTPS by omission. This allows to address a MySQL Router
bug in the OAuth flow, but is also meant to provide safer defaults.

Change-Id: I0854d6c4ac766c3c5c4a376bf6bcd776844161af