hotfix: update MRS auth exchange
This patch updates the MRS TypeScript SDK and the existing client-side
webapps (the MRS Landing Page and MRS Notes sample app) to accommodate
the changes in the MRS Auth exchange introduced by the MySQL Router.
Namely, moving POST identifying request parameters such as "authApp" and
"sessionType" to the actual request body and removing parameters that
are no longer needed.

Change-Id: I972d30c6108b02d26428ede3a61ce097564b2cc3