BUG#37371667: fix MRS TypeScript SDK null encoding
When an application uses the MRS TypeScript to retrieve REST documents
and the query filter checks for the existence of NULL or NOT NULL values
for one ore more fields, an expected error is thrown. This happens
because the underlying MRS operator values are being encoded as JSON
strings ("null") instead of JSON null, which, up until now, was ok, but
due to some breaking changes in the MySQL Router, is not anymore.

This patch updates the JSON encoding business logic used by the MRS
TypeScript SDK to ensure the MRS operator values used in those types of
requests are actual JSON null values.

Change-Id: If41c7c8e57dcf3cefcf315b158c7e710b8f14d1f