Make codicons available without editor instantiation
The Monaco Editor includes a font (codicon.ttf). However, its associated CSS, such as
.codicon-chevron-down:before { content: '\eb09'; }
is dynamically injected into the document by JavaScript via a style.monaco-colors element.
This injection only occurs if an Editor instance is created.
We require the icons to be available on pages even without an editor instance.
The only reason why it works in standalone is that CommunicationDebugger renders CodeEditor.

* Inject style.monaco-colors on application start
* Add missing monaco-editor typings definition

Change-Id: Ib4fc3332fc4b366a658b6519499afdb720fb645f