BUG#37053871, BUG#37052936 simplify SDK generation script
Generating the MRS SDK files currently requires a lot of contextual
information, which happen to be created step-by-step whilst clicking
through the corresponding buttons in the GUI. However, from an
automation standpoint, this is cumbersome, and it should be possible to
obtain that information using the data available for the existing MySQL
Shell enviroment.

One easy but huge improvement is to use the current service, as defined
by the MySQL Shell, by omission, when creating the SDK. Additionally, if
the SDK has been exported before on the same target directory, the
service id should be available in the "mrs.config.json" file. This id
can be used when the goal is to simply update the SDK code for that
service. Due a bug, this is not happening at the moment.

This patch addresses introduces the changes required to address these
two specific issues and simplifies the SDK generation script to
ultimately require only the target directory and the service URL whislt
relying on sane defaults for other configuration parameters which are
optional in nature, such as the sdk languague (which, by omission, can
be TypeScript, given it is the "native" language of the GUI).

Change-Id: I2b847b9879353ceb33232bfe71f99d092d3a4d30