BUG#37498176 use empty TypedDict for Python SDK type placeholder
Some type placeholders in the MRS Python SDK are being generated as a
None "TypeAlias". This does not work when the type is used with
"Unpack", for instance, when the dictionary fields are expanded as
keyword arguments in a function signature.

This patch introduces some changes to address this specific issue by
generating type placeholders as empty "TypedDict"s in the Python only
for the cases where the fields are actually unpacked.

Change-Id: I32654ab6c661ab856efca7c63bada65daec6029b