Fix Dropdown keyboard navigation and improve behavior
Resolve navigation and selection issues with arrow keys and Space/Enter.
Enhance handling when hovering over a dropdown item
and then using arrow keys for navigation.
Rework the Checkbox component by adding an ignoreExternalEvents prop
to allow optional state and event control from the parent component,
and reintroduce it to the DropdownItem.

Change-Id: I0aa81abda1321004e18eae35b152969c724d7868