hotfix: use POST for calling routines with the SDK
The protocol to call a REST Function or Procedure has been updated to
use an HTTP POST request instead of an HTTP PUT request.

This patch updates the MRS SDK to accommodate this change and
additionally, it re-generates the SDK to be used by the corresponding
MRS sample apps as well as the MRS Notes schema.

Change-Id: Ib8a12b345391404d64ca4a098b08d5adb67edc53