hotfix: broken pagination loop in MRS Notes
The pagination loop to retrieve the list of notes in the main page of
the MRS Notes sample app is broken, and the app keeps retrieving the
initial set of notes (based on page size) regardless of how many notes
there are.

This patch introduces a few small changes to fix that behavior.

Change-Id: I695d62e57a488f8b40e18af13de4e41da58d7b3b