BUG#37716493 avoid update conflicts in the MRS Python SDK
The MySQL Router currently ignores ETags specified using the "If-Match"
HTTP Request header. Since this is the mechanism used by the MRS Python
SDK to handle REST document update conflicts, the updates will never
fail due to a conflict, and changes that happened in the document after
it was retrieved in the first place are being overriden.

This patch adds a temporary work around to also include the ETag in the
JSON document sent in the HTTP PUT request body for an update.

Change-Id: If5c28271c97afd16099274658514c579664c6ef1