BUG#37733806: narrow sortable fields in the MRS SDK
An MRS REST View allows to specify which fields of an MRS document are
sortable or not. The generated MRS SDK code for both TypeScript and
Python does not account for the list of sortable fields when a command
allows to specify a sorting query filter and allows any existing field
to be used. This results in an HTTP 400 "Bad Request" error response
when the command is executed.

This patch extends the code generator to account for the specific list
of sortable fields which ensures developers writing an application
observe type checking errors when a sorting query filter is using fields
that are not sortable.

Change-Id: Ia7c27855d99638c9557f02df50891ca933fba734