BUG#37728029 BUG#37630549 handle update identifier
A REST View can map to a table where the primary key is composed by more
than one column. In this case, the MRS resource identifier is a
comma-separated string of the corresponding stringified object field
values. If those values do not exist, for instance, when a REST document
is retrieved beforehand and the corresponding fields are excluded from
the instance returned to the application, subsequent updates will only
incur in excess traffic.

This patch fixes an issue that prevents updates to be successful on REST
documents where the identifier is composed by mulitple fields and
introduces some logic to explicitly handle the case where those fields
do not have a value, by yielding a client-side error instead of
attempting to submit the underlying HTTP request.

Change-Id: I6599ed4143047ed7a72ef627683a28d6ab176eb7