WL#16818: MRS Python SDK - Accommodate Global Auth Apps
Authentication apps in MRS were previously tied
to a specific REST Service, at least from the
end-user standpoint. This is no longer the case,
and every authentication app is a stand-alone
element in MRS, and thus, can be linked to
multiple REST Services, and user credentials
(and any resulting access tokens) are valid between
those different services, and ultimately, even
between different applications.

With this work log, we adjust the current
authentication feature to account for the global
nature of authentication applications.

Change-Id: I3edaf52153f170648adf06693d2b7555d3a7e825