BUG#37755849 generate missing type for invisible PK
From the standpoint of MRS, invisible primary keys must always be
accounted for as REST VIEW identifiers. This is because, without them,
it is not possible for the View to be updatable.

If a table contains an invisible primary key and every other additional
column is not unique, the generated SDK incorrectly assumes there is a
binding for the primary key, when in fact, it can be excluded from the
View type alias entirely.

This patch introduces some changes to ensure that possiblity is
accounted for and all the built-in type aliases are generated, even in
the case where they should be a simple placeholder for values that
should never occur/exist.

Change-Id: Ic5118016a1b014af93a2efc7ec2ab4277ba3d4e4