WL#16891 BUG#37058508: MRS Python SDK - Support CA certificates customization for SSL
The Python SDK creates an SSL context with a default configuration,
which is quite secure, and is used when executing HTTPS requests.

This default context verifies SSL certificates for HTTPS requests,
just like a web browser. By default, SSL verification is enabled,
and Python SDK will throw an error if it’s unable to verify the
certificate (cert). A set of default “certification authority” (CA)
certificates from default locations are loaded (see SSL default certs).

Currently, there isn't a public API to customize the
configuration of the SSL context that is created, in particular,
there isn't a procedure to disable cert verification nor to specify
what certs to load if not disabled.

With this work log, the Python SDK implements an API to disable cert
verification. Also, this API allows users of the Python SDK to
specify what certs to load when cert verification is not disabled.

Change-Id: I5029494ec04a3c4d7b26fe9783d97722ed249e86