BUG#37762230 add missing sortable placeholder type
The patch to fix the issue described by BUG#37733806 introduced a
regression wherein the generated MRS Python SDK files would have type
errors if all fields of a REST object in the given service are not
sortable.

This patch adds some changes to ensure a type placeholder is generated
in that specific circumnstance.

Change-Id: I0e20159b661970de0ccabaca0ea2ae0f408bd7ea