BUG#37762456 narrow SDK list of sortable View fields
MRS allows to specify which fields of a given REST View are sortable or
not. This information is made available in the metadata schema and
should be used to narrow the list of fields that can be used to sort
documents retrieved using the MRS SDK.

Currently, the SDK allows to use any field of a REST View as part of the
order clause in a query filter. This can lead to unexpected errors if
those fields, are not specifically allowed to be sortable.

This patch introduces some changes to narrow the list of fields allowed
to be used in the order clause to the actual set of sortable fields of a
given REST View.

Change-Id: Ib164a8246a08165ff3b60e672fc2b0823b312f89