WL#16791 handle async tasks in the MRS TS SDK
REST Functions and/or Routines can take a non-trivial ammount of time to
execute. In this case, the risk of hitting a timeout in the HTTP request
or in the MySQL Router handler becomes a concern. To work around these
risky scenarios, a REST routine can be associated with an Async Task,
which allows clients to poll the MySQL Router for status updates
produced by the routine, and ultimately obtain the actual result of
executing that routine.

This patch updates the MRS TypeScript SDK, introducing a client-side
Async Task API that allows applications to execute this type of
long-running routines without having to worry about any of the issues
mentioned above.

Change-Id: I15e33b6e5e3899576e22a02a14a00f077749680d