perf: reduce db calls on the MRS SDK code generator
The MRS SDK code generator was executing a non-trivial ammount of
database queries to retrieve all the details about a given REST service,
including the list of associated REST Schemas, Objects, REST View fields
and REST Function/Procedure input parameters and results produced.

This can be a bottleneck, particularly, when in a setup with a non-local
database, where the network overhead adds up by a lot to the overhead
caused by the multiple database queries (which are already I/O bound
operations).

This patch condenses the data collection into a single MySQL Procedure
available on the mysql_rest_service_metadata schema, which is called
only once when the SDK is generated.

The metadata schema version has been bumped to 4.1.0 with the
introduction of the new Procedure.

Change-Id: I0c1a9fc8202f75e427fb9518e35b3849b3dfec1a