BUG#37755512 BUG#37766420 avoid duplicate identifiers
There are some inconsistencies when it comes to extrapolate the names
for identifiers used in the MRS SDK. Some REST resource representations,
such as Services and Objects allow to specify a custom name, whereas
REST Schemas do not. In that case, the original database object name is
used as the source.

Additionally, when it is possible to specify a custom name, there is no
validation performed and duplicate names are allowed for
representations.

As a result, when for instance, a REST Schema references the same
underlying database object, the MRS SDK will contain duplicate
identifiers.

This patch introduces some changes to ensure identifier names are
extrapolated from a unique string, such as the request path itself. This
still has a problem wherein request paths such as "/myDb" and "/my_db"
will continue to generate duplicate identifiers. In this case, the code
generator keeps track of the identifiers it generates and simply appends
a increasing suffix based on the total count of duplicates.

Change-Id: I415c4460299538f459fd5e656e2384300835c9cb