hotfix: account for undefined in async task routine
When executing a REST Routine with an associated Async Task using the
"call" command in the MRS TypeScript SDK, it is possible that the
command does not return a value, if the task has been killed while the
routine did not finish.

This patch fixes the return type of the "call" command to account for
that specific case.

Change-Id: I9d303e800f772330f71938a41698efd3c6c8adac