WL#16706: Document-level API in the MRS TS SDK
In the MRS TypeScript SDK, non-destructive REST View CRUD commands (
`create*()`, `find*()` and `update()`) always return one or more
instances of a client-side type that matches to the server-side REST
object structure. We call these instances REST documents and at the
moment they uniquely act as containers for the details of the
corresponding table/view columns in the database.

This is somewhat limiting and prevents using "shortcuts" for things
like updating the properties of the local instance and committing
those changes or even simply removing the REST Document entirely.

This patch introduces some changes in the MRS TypeScript SDK to
actually accommodate commands for this specific type of workflows.

Change-Id: I0ca5811adea41e587805b5cb10a80bf2d0b5c898