BUG#37919552 track all progress events of async task
The asynchronous generator used to produce status updates for an async
task associated to a given REST routine is not tracking the last event
produced by that task when it finishes or is cancelled. As a result of
this, when the routine is executed using the "call()" command in the MRS
TypeScript SDK, the command neither returns a result nor produces any
kind of feedback.

This patch introduces the changes required to address that issue and
make sure the generator consumer is capable of handling those specific
events and the application gets a result when it executes a REST routine
with an associated Async Task.

Change-Id: I6368c9262ccbc4724a564926a23c2c7fa3776fce