BUG#37926204 add SDK generation fallback
The MRS metadata schema 4.1.1 is missing the "EXECUTE" privileges for
the management roles assigned to plugin users. Since the SDK code
generator is now using a procedure to obtain all the required metadata
at once, the procedure will fail, when executed by a user without those
privileges.

This patch adds a workaround to ensure that the SDK metadata can still
be retrieved by the code generator when it lacks the privileges to
execute the underlying procedure. In this case, it falls back to raw
metadata schema table queries and views to retrieve individual pieces of
metadata (just like it happened on previous versions of the MRS plugin).

Change-Id: If1b2fae93b78cf919799be5eedc0178ed8d3ec3b