BUG#37914615 force lowerCamelCase for MRS TS SDK vars/props
In the MRS SDK, identifiers for class, variable and property names are
produced based on the request path of the associated REST resource. In
the case of TypeScript, the most common convention is to use
lowerCamelCase for variable and property names.

Currently, if the request path uses PascalCase, the identifier is not
converted to lowerCamelCase in this particular case.

This patch introduces some changes to ensure that lowerCamelCase is
always used when producing a variable or property name identifier in the
MRS TypeScript SDK.

Change-Id: If24a1ec42363d2d26e3396e5c5a05fb4a2fa7bf9