Fixes for SQL support for MRS roles
All usage of roles are now expected to be associated to a either a
specific service or to be global through the mrs_role.specific_to_service_id
column. ON SERVICE specifies the former and ON ANY SERVICE the latter.

Added SHOW CREATE REST ROLE and SHOW CREATE REST USER SQL statements.
Added supports for IF EXISTS on ADD/REMOVE AUTH APP clause for CREATE/ALTER
REST SERVICE.

Change-Id: I3fcaa64368114e50304dbead4884c79bfc7ba324