WL#16948 Track individual SDK version
The MRS SDK is a self-contained package that has its own requirements
in terms of versioning, particularly because there is always a
possibility to introduce public API breaking changes on one or
more of the existing language implementations in the future. Users
must be aware of the SDK version they operate.

This patch introduces some changes to the MRS SDK code generator to
ensure the `mrs.config.json` manifest file which included as part of
the generated files made available to an application, has details
about the version number of the corresponding SDK.

Change-Id: I268ef87b6e0e0129fc1a5feca8826c1affd95c40