WL#16890: MRS Python SDK - Async task API
This work log describes the changes required to introduce API
commands in the MRS Python SDK for working with REST routines
with associated async tasks.

The idea relies on producing an independent task (a different
kind of resource) when a REST routine (with an associated async
task) is called, which after starting, reports its current status
to the application. The status of the execution can be RUNNING,
CANCELLED, ERROR, TIMEOUT or COMPLETED.

If the execution terminates with a COMPLETED status, the result
(OUT parameters and result sets for the case of procedures, or
the returned value for the case of functions) of the underlying
routine is returned.

Change-Id: Ia42212db9cdbf20ca54ebc91159af41bf3a45f30