BUG#37965633 REST view data-mapping field optionality
Data-mapping fields in REST Views establish a link with a different REST
View, in the form of an underlying table or view foreign key
relationship. This allows to encode one-to-one, one-to-many and
many-to-many relationships between REST objects.

Such types of field should account for the same optionality rules
established for root-level fields when creating or updating REST
documents, when they are fully embedded as nested fields in the
top-level document, or entirely disallowed when they are expanded as new
unnested fields (since linked documents become read-only in that case).

The MRS SDK code generator is producing creatable and updatable type
aliases that do not account for these specific constraints.

This patch introduces changes to address these shortcomings.

Change-Id: I80724d36a87e9be1e35ab35dd19db82243be3f4e