WL#16855 use different iterator for multiple documents
The "findAll" and "findMany" commands in the MRS TypeScript SDK are
being replaced by a more general purpose "find" command, which
returns a custom iterable that allows to navigate between pages of
documents (as returned by the MySQL Router) without risking retrieving
insurmountable amounts of data.

This patch introduces the changes to accommodate the new API and adds
the missing "findFirstOrThrow()" command. Additionally, it also updates
the MRS sample apps with the most recent version of the SDK and the
corresponding client glue code.

Change-Id: Id53e231ad25ff88be22e46467f59c1517b21cec9