BUG#37987749 fix refresh rate validation issue
When starting a long-running routine in the MRS TypeScript SDK, it is
possible to specify custom values for the refresh rate at which status
updates produced by the associated asynchronous task are obtained, as
well as a timeout after which an appropriate status update should be
triggered if the routine did not produce a result in the meantime.

Both these options are complementary to each other and either one (or
both of them) can be omitted. However, if a timeout is specified and a
refresh rate is not, the client throws an unexpected error.

This patch addresses this issue by ensuring the vaidation of both
options happens at the correct context and point in time (i.e. before
actually starting a task). Additionally, it fixes other smaller quirk
where multiple timeout status updates are being triggered, which does
not make sense.

Change-Id: I68e6e1bd5511e237074a4ddd263e6eec012634ed