Refactor the options argument of commands Routine.call() and Routine.start()
For the sake of avoiding conflict with keyword argument
names, an arbitrary number of positional arguments are
enabled. The recommended treat is to pass only one
options dictionary, as the first positional argument.

If you, intentionally or not, pass more than one positional
argument, the option value from the later dictionary in the
sequence takes precedence

Change-Id: I3dd4c233966860704f224863445c15309cc96e0f