BUG#37716544 include etag in TS SDK document updates
In order to handle mid-air collisions, each REST document includes an
etag which must be sent by clients when the document is being updated.
This prevents clients from updating stale documents.

This patch introduces some changes in the MRS TypeScript SDK to ensure
the etag of each document, is, indeed, included in the request body sent
when the document is being updated.

Change-Id: I3c0f92ef61c147cc87733b97951a38482ff931d9