BUG#38005572 handle REST procedure with no out params
If a procedure does not specify IN/INOUT parameters, the JSON response
produced by the REST Service when calling the corresponding REST
procedure does not contain the "out_parameters" property.

In this case, if the procedure is called using the Python SDK, it
results in an unexpected error because the client is working under the
assumption that the property is always available.

This patch introduces the changes required to address this issue.

Change-Id: I8e87529fe17c615090dc52beecb379c1ec3b14ba