BUG#38043368 fix result types for Async Procedures
As described by WL#16788, the result data produced by an Asynchronous
Task associated to a REST Procedure, does not include result set data
(only the set of OUT/INOUT parameters and their corresponding values).

This particular nuance is not being accounted for by the type-safe API
commands generated for the MRS TypeScript SDK which produces a type
check error when an application tries to access the value of out
parameters produced by a REST procedure.

This patch introduces the changes required to address this issue.

Change-Id: Icf8ea7fa09126217f631c6134895865da3f43609