Handle upgrades for default heatwave endpoints
Always execute default endpoints script so that they can be updated
in case of upgrades. CREATE REST SERVICE was changed from OR REPLACE
to IF NOT EXISTS, so that the service is kept if it already exists.
Dropping the service if it exists would mean auth app links and service
specific roles would be lost in upgrades.

Change-Id: Ia3297d7fd5880403229879c824803a00400d46b3