BUG#37052890 handle MRS SDK language option
The MRS SDK is only generated for the supported programming languages if
and when the language name uses the proper convention. This is a bit too
restrictive. Additionally, the feedback provided to users when the
language is not supported is unacceptable.

Although these issues do not apply in the scope of the Shell GUI
Extension for VS Code. They are still annoying for more advanced users
that generate the MRS SDK directly using MySQL Shell commands.

This patch adds some changes that allow specifying case insensitive
programming language names when generating the MRS SDK via a MySQL Shell
command, and ensures the command raises an exception with the
appropriate message if the language is actually not supported.

Change-Id: Icaea3a7f7f1836e97043340962bbbf760cf60f99