BUG#38027483 handle SCHEDULED async tasks in the MRS SDK
An asynchronous task associated to a REST routine has an initial status
called "SCHEDULED" which is not being handled in the client SDK.
Although this status has very short duration, it is still possible,
under certain circumstances, for instance, when using the "database"
driver for the task framework, that it gains client-side exposure, and
thus, should be accounted for in order to prevent undefined behaviors.

This patch introduces the changes necessary to handle this additional
status both in the TypeScript and Python MRS SDKs.

Change-Id: Ia2701983f0c9465e57fec919cdf19abd8db0c6f0