BUG#38066815 skip REST metadata case conversion in PY SDK
MRS allows to attach JSON metadata to REST Services, Schemas and
Objects. The metadata can be retrieved by clients as-is. Currently
though, the JSON property names are being converted to snake_case in the
MRS Python SDK.

This patch introduces the changes to ensure the case conversion does not
happen.

Change-Id: I812311ec399662b0b3da8272217024606cff0b64