BUG#37750541 require filter in MRS PY SDK find_unique
The "find_unique" command in the MRS Python SDK only makes sense when a
query filter is specified for the corresponding unique fields in a REST
View. However, calling the command without any arguments does not yield
any error.

This patch introduces some changes to ensure that the query filter is
required and as a consequence, any Python type checker yields an
appropriate error when the command is called without any argument.

Change-Id: I67dfffe71f0eceac0ef33f7ed10bd0a3bc681f12