BUG#37767001 return bool in MRS SDK document level delete
The object-level "delete" command in the MRS SDK should return a boolean
value that indicates if the document was, in fact, deleted. Currently
there a few inconsistencies regarding the return value in TypeScript and
Python, and, the document-level "delete" command also does not return
any value.

This patch introduces some changes to address this inconsistency making
sure the return type of commands for delete-type operations are
consistent across all the SDK implementations. In this case, when the
command is supposed to delete a single document, the return value is a
boolean. On the other hand, if the command is supposed to return or or
more documents, the return value is a numeric indicating the number of
documents that have been deleted.

Change-Id: Ice4d16bdc8341bbd5476202a6711df8772c4a984