BUG#38248624 use consistent auth property names in the MRS SDK
Authentication parameters should follow the same convention on all
flavors of the MRS SDK. Currently this is not the case, and there is a
mix of incorrect documentation examples and actual inconsistencies in
the Python implementation.

This patch establishes an appropriate convention for authentication
parameters and introduces some changes to enforce that convention both
from the documentation standpoint as well as the implementation
standpoint.

Additionally, it also addresses some loosely related documentation
incorrections.

Change-Id: I5bebb30561e13402ba84bb784df6e22f1b882162