BUG#38257852 fix REST function SDK param and result mapping
REST Function parameter and result field mapping in the MRS SDK breaks
down on different platforms due to potential different order in the
result sets produced by the MySQL server when retrieving the details of
REST objects from the metadata schema.

This is the result of a limitation in the sdk_service_data procedure
which is not having into account SDK object positions in the order
clause, contrary to what happens when objects are retrieved in
individual database calls.

This introduces some changes to account for this circumstance. Namely,
it upgrades the MRS metadata schema with the new version of the
sdk_service_data procedure and adds a corresponding business logic
failsafe to avoid using magic numbers.

Change-Id: Ie4cd67dca5da1f0ab5079d7250f917e83caba3db